1. The API key in the website for the feed is fully functional - it no longer comes from the service config.
2. Waiting messages when a feed is being created or updated.

3. Notification messages after a successful create or update.

4. Performance improvements when starting up multiple feeds.

5. Paging on the feeds page

+++ src/NuFridge.Service/Website/Controllers/FeedsController.cs
@@ -1,7 +1,12 @@
-﻿using System.Net;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Net;
 using System.Net.Http;
 using System.Web.Http;
 using System.Web.Http.OData;
+using System.Web.Http.Routing;
+using NuFridge.Service.Feeds;
 using NuFridge.Service.Model;
 using NuFridge.Service.Repositories;
 
@@ -17,14 +22,30 @@ namespace NuFridge.Service.Website.Controllers
             FeedRepository = new SqlCompactRepository<Feed>();
         }
 
-        [EnableQuery]
-        public HttpResponseMessage Get(string id = "")
+        public Object Get(int page = 0, int pageSize = 5)
         {
-            if (id == string.Empty)
-            {
-                return Request.CreateResponse(FeedRepository.GetAll());
-            }
+            IList<Feed> query = FeedRepository.GetAll();
 
+            var totalCount = query.Count();
+            var totalPages = (int)Math.Ceiling((double)totalCount / pageSize);
+
+            var urlHelper = new UrlHelper(Request);
+
+            var results = query
+                .Skip(pageSize*page)
+                .Take(pageSize)
+                .ToList();
+
+            return new
+            {
+                TotalCount = totalCount,
+                TotalPages = totalPages,
+                Results = results
+            };
+        }
+
+        public HttpResponseMessage Get(string id)
+        {
             var feed = FeedRepository.GetById(id);
 
             if (feed == null)
@@ -40,9 +61,34 @@ namespace NuFridge.Service.Website.Controllers
             return Request.CreateResponse(feed);
         }
 
+        //[EnableQuery]
+        //public HttpResponseMessage Get(string id = "")
+        //{
+        //    if (id == string.Empty)
+        //    {
+        //        return Request.CreateResponse(FeedRepository.GetAll());
+        //    }
+
+        //    var feed = FeedRepository.GetById(id);
+
+        //    if (feed == null)
+        //    {
+        //        var resp = new HttpResponseMessage(HttpStatusCode.NotFound)
+        //        {
+        //            Content = new StringContent(string.Format("No feed with ID = {0}", id)),
+        //            ReasonPhrase = "Feed not found"
+        //        };
+        //        throw new HttpResponseException(resp);
+        //    }
+
+        //    return Request.CreateResponse(feed);
+        //}
+
         public HttpResponseMessage Post([FromBody]Feed feed)
         {
             FeedRepository.Insert(feed);
+
+            FeedManager.Instance().Start(feed);
 
             return Request.CreateResponse(HttpStatusCode.OK, feed);
         }
@@ -54,7 +100,11 @@ namespace NuFridge.Service.Website.Controllers
                 return Request.CreateErrorResponse(HttpStatusCode.BadRequest, "Feed id does not match.");
             }
 
+            FeedManager.Instance().Stop(feed);
+
             FeedRepository.Update(feed);
+
+            FeedManager.Instance().Start(feed);
 
             return Request.CreateResponse(HttpStatusCode.OK, feed);
         }
@@ -73,9 +123,21 @@ namespace NuFridge.Service.Website.Controllers
                 throw new HttpResponseException(resp);
             }
 
+            var success = FeedManager.Instance().Stop(feed);
+
+            if (!success)
+            {
+                var resp = new HttpResponseMessage(HttpStatusCode.InternalServerError)
+                {
+                    Content = new StringContent(string.Format("Failed to stop feed with ID = {0}", id)),
+                    ReasonPhrase = "Failed to stop feed."
+                };
+                throw new HttpResponseException(resp);
+            }
+
             FeedRepository.Delete(feed);
 
-            return Request.CreateResponse(HttpStatusCode.OK);
+            return Request.CreateResponse(HttpStatusCode.OK, feed);
         } 
     }
 }
